+++ src/models.ts
@@ -28,6 +28,7 @@ export interface Settings {
   bottomOffset: number,
   darkMode: boolean;
   bottomClose: boolean;
+  fastSwipeClose: boolean;
   freeMode: boolean;
   buttonClose: boolean;
   topperOverflow: boolean;